+++ CHANGELOG.md
@@ -7,6 +7,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Fixed
+- Display required external input names instead of `[object Object]` in prover errors
+
 ## [2.0.11] - 2025-09-30
 
 ### Changed